+++ spec/features/users_can_view_the_site_navigation_spec.rb
@@ -19,13 +19,13 @@ RSpec.feature "Users can view the site navigation" do
     before { authenticate!(user: user) }
     after { logout }
 
-    it "shows the appropriate naviation" do
+    it "shows the appropriate navigation" do
       visit organisation_path(user.organisation)
 
       expect(page).to have_css ".govuk-header__navigation"
       expect(page).to have_link t("page_title.home"), href: home_path
       expect(page).to have_link t("page_title.report.index"), href: reports_path
-      expect(page).not_to have_link t("page_title.organisation.index"), href: organisations_path, class: "govuk-header__link"
+      expect(page).to have_link t("page_title.organisation.index"), href: organisations_path, class: "govuk-header__link"
       expect(page).not_to have_link t("page_title.users.index"), href: users_path
     end
   end
@@ -35,7 +35,7 @@ RSpec.feature "Users can view the site navigation" do
     before { authenticate!(user: user) }
     after { logout }
 
-    it "shows the appropriate naviation" do
+    it "shows the appropriate navigation" do
       visit organisation_path(user.organisation)
 
       expect(page).to have_css ".govuk-header__navigation"